Move chunk definition parsing into CLI parser
